pyproject.toml
[project]
name = "appeals-backend"
version = "0.1.0"
description = "Minimal-viable prototype of a citizen-appeals processing system: FastAPI REST backend + async Telegram bot, ready for further extension and production hardening."
readme = "readme.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "pydantic>=2.11.5",
    "tortoise-orm>=0.25.0",
    "uvicorn>=0.34.2",
]

[tool.pyright]
venvPath = '.'
venv = '.venv'
